Factor out `harmonia-store-nar-info` crate
Extract NarInfo construction and text formatting from `harmonia-cache`
into a dedicated crate. Also adds `UnkeyedNarInfo` with serde support
matching the upstream Nix JSON format (version 3), with upstream JSON
tests for both pure and impure variants.

Introduce `StorePathKeyed<T>` generic wrapper in `harmonia-store-path-info`
so that `ValidPathInfo` becomes `StorePathKeyed<UnkeyedValidPathInfo>` and
`NarInfo` becomes `StorePathKeyed<UnkeyedNarInfo>`.

Fix the `?json=1` endpoint in `harmonia-cache` to serialize the full
nar-info JSON format (with `url`, `compression`, `downloadHash`,
`downloadSize`) instead of just the bare path-info.

Simplify `json_upstream` test layout in `harmonia-store-path-info` and
`harmonia-store-build-result` to use a single module file instead of a
separate submodule per type.
