+++ Cargo.toml
@@ -7,6 +7,7 @@ members = [
     "harmonia-utils-hash",
     "harmonia-store-core",
     "harmonia-store-db",
+    "harmonia-store-nar-info",
     "harmonia-store-path-info",
     "harmonia-store-remote",
     "harmonia-store-aterm",
@@ -99,6 +100,7 @@ harmonia-store-aterm         = { path = "harmonia-store-aterm" }
 harmonia-store-build-result  = { path = "harmonia-store-build-result" }
 harmonia-store-core          = { path = "harmonia-store-core" }
 harmonia-store-db            = { path = "harmonia-store-db" }
+harmonia-store-nar-info      = { path = "harmonia-store-nar-info" }
 harmonia-store-remote        = { path = "harmonia-store-remote" }
 harmonia-utils-base-encoding = { path = "harmonia-utils-base-encoding" }
 harmonia-utils-hash          = { path = "harmonia-utils-hash" }
