+++ harmonia-cache/Cargo.toml
@@ -45,6 +45,7 @@ futures-util             = { workspace = true }
 harmonia-nar             = { version = "3.0.0" }
 harmonia-store-core      = { version = "0.0.0-alpha.0" }
 harmonia-store-db        = { version = "3.0.0" }
+harmonia-store-nar-info  = { path = "../harmonia-store-nar-info" }
 harmonia-store-path-info = { path = "../harmonia-store-path-info" }
 harmonia-utils-hash      = { version = "0.0.0-alpha.0" }
 nix                      = { workspace = true, features = [ "socket", "fs" ] }
