+++ harmonia-store-nar-info/Cargo.toml
@@ -1,0 +1,29 @@
+[package]
+name                 = "harmonia-store-nar-info"
+version.workspace    = true
+authors.workspace    = true
+edition.workspace    = true
+license.workspace    = true
+homepage.workspace   = true
+repository.workspace = true
+description          = "NarInfo formatting and construction for Nix binary cache protocol"
+
+[dependencies]
+harmonia-store-core      = { path = "../harmonia-store-core" }
+harmonia-store-path-info = { path = "../harmonia-store-path-info" }
+harmonia-utils-hash      = { version = "0.0.0-alpha.0" }
+serde                    = { workspace = true }
+
+[dev-dependencies]
+harmonia-store-core      = { path = "../harmonia-store-core" }
+harmonia-store-path-info = { path = "../harmonia-store-path-info" }
+harmonia-utils-test      = { version = "0.0.0-alpha.0" }
+hex-literal              = { workspace = true }
+serde_json               = { workspace = true }
+
+[[test]]
+name = "json_upstream"
+path = "tests/json_upstream/mod.rs"
+
+[lints]
+workspace = true
